requirements.txt
Flask>=2.3.0
tensorflow>=2.15.0
opencv-python>=4.8.1.78
pillow>=10.0.0
numpy>=1.26.0
pandas>=2.1.0
scikit-learn>=1.3.0
matplotlib>=3.8.0
seaborn>=0.13.0
h5py>=3.9.0
streamlit>=1.31.0


